+++ onsiteprint-plugin/blocks/event/block-script-parts/parts.js
@@ -1,19 +1,21 @@
 /* ------------------------------------------------------------------------
  #  JS Part Name: Parts
  *  Block functions included in the Event Block Script.
- ?  Updated: 2025-07-03 - 01:31 (Y:m:d - H:i)
- ?  Info: Added new Script, participant-listeners.js.
+ ?  Updated: 2025-12-14 - 05:28 (Y:m:d - H:i)
+ ?  Info: Added new Script, setup-footer.js.
 ---------------------------------------------------------------------------
  #  TABLE OF CONTENTS:
 ---------------------------------------------------------------------------
 
     1. 	Setup Header
     
-    2. 	Setup List
+    2. 	Setup Footer
 
-    3.  Add Participant
+    3. 	Setup List
 
-    4.  Participant Listeners
+    4.  Add Participant
+
+    5.  Participant Listeners
 
 ---------------------------------------------------------------------------
  #  1. Setup Header
@@ -22,19 +24,25 @@ import { opSetupHeader } from './setup-header.js'
 export { opSetupHeader }
 
 /* ------------------------------------------------------------------------
- #  2. Setup List
+ #  2. Setup Footer
+--------------------------------------------------------------------------- */
+import { opSetupFooter } from './setup-footer.js'
+export { opSetupFooter }
+
+/* ------------------------------------------------------------------------
+ #  3. Setup List
 --------------------------------------------------------------------------- */
 import { opSetupList } from './setup-list.js'
 export { opSetupList }
 
 /* ------------------------------------------------------------------------
- #  3. Add Participant
+ #  4. Add Participant
 --------------------------------------------------------------------------- */
 import { opAddParticipant } from './add-participant.js'
 export { opAddParticipant }
 
 /* ------------------------------------------------------------------------
- #  4. Participant Listeners
+ #  5. Participant Listeners
 --------------------------------------------------------------------------- */
 import { opParticipantToggleListener, opPrintParticipantListener } from './participant-listeners.js'
 export {  opParticipantToggleListener, opPrintParticipantListener }